Change config RAM handling to be significantly smaller.
This changes the default for SystemConfig to be (1) a const and (2)
all-zeroes when represented in memory. That means we can just initialize
the static config RAM directly, instead of doing a dance with
MaybeUninit. And the actual initialization of the RAM now falls to the
general BSS setup loop in the runtime, which we were already paying for.